Rework on prelude by multi-level design
