+++ src/tests/binary.spec.ts
@@ -1,6 +1,11 @@
 import test from 'ava';
+import crypto from 'crypto';
+import fs from 'fs';
+import path from 'path';
 import shell from 'shelljs';
-import { compileToLLVMIR } from './util';
+
+import LLVMCodeGen from '../codegen';
+import Prelude from '../prelude';
 
 const srcArgsFromShell = `
 function main(argc: number, argv: string[]): number {
@@ -12,13 +17,26 @@ function main(argc: number, argv: string[]): number {
 `;
 
 test('test args', async t => {
-  const irPath = await compileToLLVMIR(srcArgsFromShell);
-  const name = irPath.slice(0, irPath.length - 3);
+  const hash = crypto
+    .createHash('md5')
+    .update(srcArgsFromShell)
+    .digest()
+    .toString('hex');
 
-  shell.exec(`llvm-as ${name}.ll -o ${name}.bc`, { async: false });
-  shell.exec(`llc -filetype=obj ${name}.bc -o ${name}.o`, { async: false });
-  shell.exec(`gcc ${name}.o -o ${name}`, { async: false });
-  const ret = shell.exec(`${name} foo bar`, { async: false });
+  fs.mkdirSync(path.join(shell.tempdir(), 'minits'), { recursive: true });
+  const name = path.join(shell.tempdir(), 'minits', hash + '.ts');
+  fs.writeFileSync(name, srcArgsFromShell);
+
+  const prelude = new Prelude(name);
+  const outputs = prelude.process();
+  const codegen = new LLVMCodeGen([outputs]);
+  codegen.genSourceFile(outputs);
+
+  const full = path.join(path.dirname(outputs), 'output');
+  shell.exec(`llvm-as ${full}.ll -o ${full}.bc`, { async: false });
+  shell.exec(`llc -filetype=obj ${full}.bc -o ${full}.o`, { async: false });
+  shell.exec(`gcc ${full}.o -o ${full}`, { async: false });
+  const ret = shell.exec(`${full} foo bar`, { async: false });
   t.log(ret.stdout);
   if (ret.code === 0) {
     t.pass();
